fix expansion of types that have multiple layers of aliasing
Types like :bool, which aliases :boolean, which aliases a primitive
type like :char.

Care is taken to call the right expansion: direct or indirect.

As a result, the special case for :enum is no longer needed, and the
fsbv.makepair tests are no longer expected to fail.

There is still the issue where an indirect expansion may resort to
returning an unsuitable (direct) body.  Perhaps the protocol should be
made less brittle in order to facilitate a complete solution.